Refactor tests to their own directory
