+++ Cargo.toml
@@ -14,3 +14,6 @@ categories = ["memory-management", "no-std"]
 experimental-ghost-cursor = []
 # Enables the use of unproven multiple mutable borrows.
 experimental-multiple-mutable-borrows = []
+
+[dev-dependencies]
+trybuild = "1.0"
