+++ src/ghost_cell.rs
@@ -356,113 +356,3 @@ unsafe impl<'brand, T: ?Sized + Send + Sync> Sync for GhostCell<'brand, T> {}
 //
 
 type InvariantLifetime<'brand> = PhantomData<fn(&'brand ()) -> &'brand ()>;
-
-#[doc(hidden)]
-pub mod compile_tests {
-
-/// ```compile_fail
-/// use ghost_cell::{GhostToken, GhostCell};
-///
-/// GhostToken::new(|token| token);
-/// ```
-pub fn token_noescape() {}
-
-/// ```compile_fail
-/// use ghost_cell::{GhostToken, GhostCell};
-///
-/// GhostToken::new(|mut token| {
-///     let cell = GhostCell::new(42);
-///
-///     *cell.borrow_mut(&mut token) = 33;
-///
-///     cell
-/// });
-/// ```
-pub fn cell_noescape() {}
-
-/// ```compile_fail,E0505
-/// use ghost_cell::{GhostToken, GhostCell};
-///
-/// GhostToken::new(|token| {
-///     let cell = GhostCell::new(42);
-///
-///     let r = cell.borrow(&token);
-///     std::mem::drop(token);
-///
-///     *r
-/// });
-/// ```
-pub fn cell_borrow_borrows_token() {}
-
-/// ```compile_fail,E0502
-/// use ghost_cell::{GhostToken, GhostCell};
-///
-/// GhostToken::new(|mut token| {
-///     let one = GhostCell::new(1);
-///     let two = GhostCell::new(2);
-///
-///     let r = one.borrow_mut(&mut token);
-///     assert_eq!(2, *two.borrow(&token));
-///
-///     *r = 33;
-/// });
-/// ```
-pub fn cell_borrow_mut_borrows_token_mutably() {}
-
-/// ```compile_fail,E0505
-/// use ghost_cell::{GhostToken, GhostCell};
-///
-/// GhostToken::new(|token| {
-///     let cell = GhostCell::new(42);
-///
-///     let r = cell.borrow(&token);
-///     std::mem::drop(cell);
-///
-///     *r
-/// });
-/// ```
-pub fn cell_borrow_borrows_cell() {}
-
-/// ```compile_fail,E0505
-/// use ghost_cell::{GhostToken, GhostCell};
-///
-/// GhostToken::new(|mut token| {
-///     let cell = GhostCell::new(42);
-///
-///     let r = cell.borrow_mut(&mut token);
-///     std::mem::drop(cell);
-///
-///     *r
-/// });
-/// ```
-pub fn cell_borrow_mut_borrows_cell() {}
-
-/// ```compile_fail,E0502
-/// use ghost_cell::{GhostToken, GhostCell};
-///
-/// GhostToken::new(|token| {
-///     let mut cell = GhostCell::new(42);
-///
-///     let r = cell.get_mut();
-///     assert_eq!(42, *cell.borrow(&token));
-///
-///     *r = 33;
-/// });
-/// ```
-pub fn cell_get_mut_borrows_cell_mutably() {}
-
-/// ```compile_fail,E0502
-/// use ghost_cell::{GhostToken, GhostCell};
-///
-/// GhostToken::new(|token| {
-///     let mut value = 42;
-///
-///     let cell = GhostCell::from_mut(&mut value);
-///
-///     assert_eq!(42, value);
-///     assert_eq!(42, *cell.borrow(&token));
-/// });
-/// ```
-pub fn cell_from_mut_borrows_value_mutably() {}
-
-} // mod compile_tests
