+++ src/ghost_cursor.rs
@@ -268,88 +268,3 @@ unsafe fn as_ref<'a, T: ?Sized>(ptr: NonNull<T>) -> &'a T {
 unsafe fn as_mut<'a, T: ?Sized>(ptr: NonNull<T>) -> &'a mut T {
     &mut *ptr.as_ptr()
 }
-
-#[doc(hidden)]
-pub mod compile_tests {
-
-/// ```compile_fail,E0502
-/// use ghost_cell::{GhostCell, GhostCursor, GhostToken};
-///
-/// GhostToken::new(|mut token| {
-///     let cell = GhostCell::new(1);
-///     let cursor = GhostCursor::new(&mut token, Some(&cell));
-///
-///     assert_eq!(1, *cell.borrow(&token));
-///     assert_eq!(Some(&1), cursor.borrow());
-/// })
-/// ```
-pub fn cursor_new_borrows_token_mutably() {}
-
-/// ```compile_fail,E0502
-/// use ghost_cell::{GhostCell, GhostCursor, GhostToken};
-///
-/// GhostToken::new(|mut token| {
-///     let (one, two) = (GhostCell::new(1), GhostCell::new(2));
-///
-///     let cursor = GhostCursor::new(&mut token, Some(&one));
-///     if let Some(one) = cursor.into_inner() {
-///         assert_eq!(2, *two.borrow(&token));  //  Fail, token still borrowed by `one`.
-///         *one = 3;
-///     }
-/// })
-/// ```
-pub fn cursor_into_inner_leaves_token_borrowed_mutably() {}
-
-/// ```compile_fail,E0499
-/// use ghost_cell::{GhostCell, GhostCursor, GhostToken};
-///
-/// GhostToken::new(|mut token| {
-///     let (one, two) = (GhostCell::new(1), GhostCell::new(2));
-///
-///     let cursor = GhostCursor::new(&mut token, Some(&one));
-///     let aliased_token = cursor.into_parts().0;
-///     *two.borrow_mut(&mut token) = 4;   //  Fail, token still borrowed by `aliased_token`.
-///     assert_eq!(1, *one.borrow(aliased_token));
-/// })
-/// ```
-pub fn cursor_into_parts_first_part_leaves_token_borrowed_mutably() {}
-
-/// ```compile_fail,E0502
-/// use ghost_cell::{GhostCell, GhostCursor, GhostToken};
-///
-/// GhostToken::new(|mut token| {
-///     let (one, two) = (GhostCell::new(1), GhostCell::new(2));
-///
-///     let cursor = GhostCursor::new(&mut token, Some(&one));
-///     if let Some(one) = cursor.into_parts().1 {
-///         *two.borrow_mut(&mut token) = 4;   //  Fail, token still borrowed by `one`.
-///         assert_eq!(1, *one.borrow(&token));
-///     }
-/// })
-/// ```
-pub fn cursor_into_parts_second_part_leaves_token_borrowed_mutably() {}
-
-/// ```compile_fail,E0521
-/// use core::cell::Cell;
-/// use ghost_cell::{GhostCell, GhostCursor, GhostToken};
-///
-/// GhostToken::new(|mut token| {
-///     let cell = GhostCell::new(1);
-///     let leak = Cell::new(None);
-///     let mut cursor = GhostCursor::new(&mut token, Some(&cell));
-///
-///     let _ = cursor.move_mut(|cell_ref| {
-///         leak.set(Some(cell_ref));   //  Fail, `cell_ref` cannot escape the closure body.
-///         None
-///     });
-///
-///     //  If `cell_ref` escaped, this would be a shared reference whose value can change -- this is unsound.
-///     let cell_ref: &i32 = leak.get().unwrap();
-///     assert_eq!(*cell_ref, 1);
-///     *cursor.borrow_mut().unwrap() = 42;
-///     assert_eq!(*cell_ref, 42);
-/// })
-/// ```
-pub fn cursor_move_mut_noescape() {}
-
-} // mod compile_tests
