+++ tests/compile_tests.rs
@@ -1,0 +1,7 @@
+#[test]
+fn compile_tests() {
+    let t = trybuild::TestCases::new();
+    t.compile_fail("tests/compile_tests/ghost_cell/*.rs");
+    #[cfg(feature = "experimental-ghost-cursor")]
+    t.compile_fail("tests/compile_tests/ghost_cursor/*.rs");
+}
