+++ tests/compile_tests/ghost_cell/cell_borrow_borrows_token.rs
@@ -1,0 +1,12 @@
+use ghost_cell::{GhostCell, GhostToken};
+
+fn main() {
+    GhostToken::new(|token| {
+        let cell = GhostCell::new(42);
+
+        let r = cell.borrow(&token);
+        std::mem::drop(token);
+
+        *r
+    });
+}
