+++ tests/compile_tests/ghost_cell/cell_borrow_borrows_token.stderr
@@ -1,0 +1,10 @@
+error[E0505]: cannot move out of `token` because it is borrowed
+  --> tests/compile_tests/ghost_cell/cell_borrow_borrows_token.rs:8:24
+   |
+7  |         let r = cell.borrow(&token);
+   |                             ------ borrow of `token` occurs here
+8  |         std::mem::drop(token);
+   |                        ^^^^^ move out of `token` occurs here
+9  |
+10 |         *r
+   |         -- borrow later used here
