+++ tests/compile_tests/ghost_cell/cell_borrow_mut_borrows_token_mutably.rs
@@ -1,0 +1,13 @@
+use ghost_cell::{GhostCell, GhostToken};
+
+fn main() {
+    GhostToken::new(|mut token| {
+        let one = GhostCell::new(1);
+        let two = GhostCell::new(2);
+
+        let r = one.borrow_mut(&mut token);
+        assert_eq!(2, *two.borrow(&token));
+
+        *r = 33;
+    });
+}
