+++ tests/compile_tests/ghost_cell/cell_from_mut_borrows_value_mutably.rs
@@ -1,0 +1,12 @@
+use ghost_cell::{GhostCell, GhostToken};
+
+fn main() {
+    GhostToken::new(|token| {
+        let mut value = 42;
+
+        let cell = GhostCell::from_mut(&mut value);
+
+        assert_eq!(42, value);
+        assert_eq!(42, *cell.borrow(&token));
+    });
+}
