+++ tests/compile_tests/ghost_cell/cell_get_mut_borrows_cell_mutably.rs
@@ -1,0 +1,12 @@
+use ghost_cell::{GhostCell, GhostToken};
+
+fn main() {
+    GhostToken::new(|token| {
+        let mut cell = GhostCell::new(42);
+
+        let r = cell.get_mut();
+        assert_eq!(42, *cell.borrow(&token));
+
+        *r = 33;
+    });
+}
