+++ tests/compile_tests/ghost_cell/cell_get_mut_borrows_cell_mutably.stderr
@@ -1,0 +1,10 @@
+error[E0502]: cannot borrow `cell` as immutable because it is also borrowed as mutable
+  --> tests/compile_tests/ghost_cell/cell_get_mut_borrows_cell_mutably.rs:8:25
+   |
+7  |         let r = cell.get_mut();
+   |                 -------------- mutable borrow occurs here
+8  |         assert_eq!(42, *cell.borrow(&token));
+   |                         ^^^^^^^^^^^^^^^^^^^ immutable borrow occurs here
+9  |
+10 |         *r = 33;
+   |         ------- mutable borrow later used here
