+++ tests/compile_tests/ghost_cell/cell_noescape.stderr
@@ -1,0 +1,14 @@
+error: lifetime may not live long enough
+ --> tests/compile_tests/ghost_cell/cell_noescape.rs:7:9
+  |
+4 |     GhostToken::new(|mut token| {
+  |                      ---------- return type of closure is GhostCell<'2, i32>
+  |                      |
+  |                      has type `GhostToken<'1>`
+...
+7 |         cell
+  |         ^^^^ returning this value requires that `'1` must outlive `'2`
+  |
+  = note: requirement occurs because of the type `GhostCell<'_, i32>`, which makes the generic argument `'_` invariant
+  = note: the struct `GhostCell<'brand, T>` is invariant over the parameter `'brand`
+  = help: see <https://doc.rust-lang.org/nomicon/subtyping.html> for more information about variance
