+++ tests/compile_tests/ghost_cell/token_noescape.rs
@@ -1,0 +1,5 @@
+use ghost_cell::GhostToken;
+
+fn main() {
+    GhostToken::new(|token| token);
+}
