+++ tests/compile_tests/ghost_cell/token_noescape.stderr
@@ -1,0 +1,12 @@
+error: lifetime may not live long enough
+ --> tests/compile_tests/ghost_cell/token_noescape.rs:4:29
+  |
+4 |     GhostToken::new(|token| token);
+  |                      ------ ^^^^^ returning this value requires that `'1` must outlive `'2`
+  |                      |    |
+  |                      |    return type of closure is GhostToken<'2>
+  |                      has type `GhostToken<'1>`
+  |
+  = note: requirement occurs because of the type `GhostToken<'_>`, which makes the generic argument `'_` invariant
+  = note: the struct `GhostToken<'brand>` is invariant over the parameter `'brand`
+  = help: see <https://doc.rust-lang.org/nomicon/subtyping.html> for more information about variance
